requirements.txt
fastapi
uvicorn
pillow
torch
torchvision
tensorflow
gradio

